pyproject.toml
[project]
name = "kagami"
version = "0.1.0"
description = "GitHub bot with Claude Code integration"
requires-python = ">=3.11"
dependencies = [
    "PyGithub>=2.1.1",
    "playwright>=1.56.0",
    "playwright-stealth>=2.0.0",
    "proxy.py>=2.4.0",  # HTTP/2対応のプロキシライブラリ
    "httpx[http2]>=0.27.0",  # HTTP/2クライアント（h2パッケージ込み）
    "selenium>=4.38.0",  # ブラウザ自動化（Playwrightの代替）
    "undetected-chromedriver>=3.5.5",
    "webdriver-manager>=4.0.2",  # ChromeDriverの自動管理
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.uv]
dev-dependencies = []

# スクリプトベースのプロジェクトなので、ビルドバックエンドは不要
